[prog] enable by default `lsp-bridge' as the defualt lsp client.
lsp-bridge is yet another masterpiece from lazycat. instead of
optimizing emacs-side code, it moves lsp interaction from emacs to
another server process, and thus the interaction becomes fully
*asynchronous*. the lsp-bridge server (not lsp server) decides whether
or not to display a popup completion window, distinctly contrast to
the traditional `complete-at-point' way.

the older way:

1. put your point on a symbol.
2. a timer triggers capf.
3. capf sends a completion request to lsp client.
4. lsp client *synchronously* prepares a lsp rpc call and waits for
   the response.

lsp-bridge:

1. it actively tracks the cursor in the buffer (the lsp-bridge client
   sends such information to the lsp-bridge server.)
2. the lsp-bridge server interacts with lsp servers _in the
   background_. it sends back any responses.
3. the lsp-bridge server controls when corfu shows.

the technology is not surprising at all. and... actually it can be
used for a whole lot more scenarios! this is what interests me most.
i intend to dig deeper into all this stuff...