[ui] disable tab-bar for now
tab-bar is not as useful as I thought.

The primary reason I enabled it is I want to put several "frames"
together.  But tab-bar can get in the way of winner, so it is worse
than frames.

Frames are also bad, because global states are shared.  I wish I can
optimize the startup time and move to a separate-process solution.