[completion] Use company instead of corfu.
Corfu is lacking in some important ways. For example, it is difficult
to get corfu to use Dabbrev as a completion backend. It is not
impossible, but it is harder than necessary.  OTOH, Company has OOTB
support for Dabbrev completion.

One of the more significant problems is that Corfu has poor
integration with Eglot (or the other way around: Eglot has poor
integration with Corfu). I'm not interested in finding out who is to
blame, but only interested in *getting a working environment*. Because
it is easy to swap out Corfu, and Company already has an established
ecosystem, I decided to let Corfu go.